requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Image processing
Pillow>=9.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP requests
requests>=2.31.0

# Progress bars
tqdm>=4.65.0

# CLIP model
clip-by-openai>=1.0

# BLIP model (included in transformers)
# transformers already includes BLIP models

# Optional: for better performance
# accelerate>=0.20.0
# bitsandbytes>=0.39.0
